requirements.txt
Flask>=2.3.0
line-bot-sdk>=3.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
redis>=4.0.0
gunicorn>=20.0.0